Able to sync with jellyfin | Slight improvements

+++ Program.cs
@@ -1,14 +1,26 @@
 using Microsoft.AspNetCore.Mvc;
 
+using static System.Net.WebRequestMethods;
+
 //TODO: Configuration & Secrets
-const string APIKEY = "MTY4NjYxNTAzMTI4NjQ0OGY2ODdlLTRlMGYtNGEyMC1iY2EwLWI0OWQ3NDkzOGRhNg==";
+const string JELLYSEERR_APIKEY = "MTY4NjYxNTAzMTI4NjQ0OGY2ODdlLTRlMGYtNGEyMC1iY2EwLWI0OWQ3NDkzOGRhNg==";
+const string JELLYFIN_APIKEY = "05092613243f40039715afb2affc8eb3";
+
+const string JELLYFIN_HOST_URL = "http://192.168.1.10:8096/";
+const string JELLYSEERR_HOST_URL = "http://192.168.1.11:5055/";
 
 var builder = WebApplication.CreateBuilder( args );
 
-builder.Services.AddHttpClient( "JellySeerr", ( client ) =>
+builder.Services.AddHttpClient( "Jellyseerr", ( client ) =>
 {
-    client.BaseAddress = new Uri( "http://192.168.1.11:5055/api/v1/" );
-    client.DefaultRequestHeaders.Add( "X-Api-Key", APIKEY );
+    client.BaseAddress = new Uri( $"{JELLYSEERR_HOST_URL}/api/v1/" );
+    client.DefaultRequestHeaders.Add( "X-Api-Key", JELLYSEERR_APIKEY );
+} );
+
+builder.Services.AddHttpClient( "Jellyfin", ( client ) =>
+{
+    client.BaseAddress = new Uri( JELLYFIN_HOST_URL );
+    client.DefaultRequestHeaders.Add( "Authorization", $"MediaBrowser Token=\"{JELLYFIN_APIKEY}\"" );
 } );
 
 var app = builder.Build();
@@ -17,17 +29,58 @@ var app = builder.Build();
 
 app.UseHttpsRedirection();
 
+app.MapGet( "/", async ( context ) => await context.Response.WriteAsync( @"Hello World!
 
-app.MapGet( "test", async ( [FromServices] IHttpClientFactory httpClientFactory ) =>
+Endpoints:
+/radarr/notification
+/sonarr/notification
+" ) );
+
+app.MapGet( "/sync", async ( [FromServices] IHttpClientFactory httpClientFactory ) =>
 {
-    var client = httpClientFactory.CreateClient( "JellySeerr" );
-    var movie = await client.GetFromJsonAsync<JellySeerrMovie>( $"movie/{910571}" );
+    Console.WriteLine( "Processing Sync..." );
 
-    if (movie?.MediaInfo is not null)
+    var jellyfinClient = httpClientFactory.CreateClient( "Jellyfin" );
+    var jellyseerrClient = httpClientFactory.CreateClient( "Jellyseerr" );
+
+    var searchResult = await jellyseerrClient.GetFromJsonAsync<JellyseerrMediaSearchResult>( $"media?take=999&skip=0&filter=available&sort=added" );
+    var availableMovies = searchResult.Results.Where( x => x.MediaType.Equals( "movie", StringComparison.InvariantCultureIgnoreCase ) );
+    
+
+    var moviesIds = availableMovies
+        .Where( x => !string.IsNullOrWhiteSpace( x.JellyfinMediaId ) || !string.IsNullOrWhiteSpace( x.JellyfinMediaId4k ) )
+        .Select( x => new
+        {
+            Id = string.IsNullOrWhiteSpace( x.JellyfinMediaId )
+        ? Guid.Parse(x.JellyfinMediaId4k).ToString( "d" )
+        : Guid.Parse(x.JellyfinMediaId).ToString("d"),
+            MediaId = x.Id,
+            TmdbId = x.TmdbId
+        } );
+
+    Console.WriteLine( "Available Jellyfin Movies: " + moviesIds.Count() );
+
+
+    var jellySearchResult = await jellyfinClient.GetFromJsonAsync<JellyfinSearchResult>( $"Items?ids={string.Join(",", moviesIds.Select(x=> x.Id))}&enableTotalRecordCount=false&enableImages=false" );
+
+    var notFoundMovies = moviesIds.Where( x => !jellySearchResult.Items.Any( y => Guid.Parse(y.Id) == Guid.Parse(x.Id) ) );
+    var links = notFoundMovies.Select( x => $"{JELLYSEERR_HOST_URL}/movie/{x.TmdbId}" );
+
+    Console.WriteLine( "Not Found Movies To Sync: " + notFoundMovies.Count() );
+
+    //foreach (var link in links)
+    //{
+    //    Console.WriteLine( link );
+    //}
+
+    if (notFoundMovies.Count() > 0)
     {
-        Console.WriteLine( System.Text.Json.JsonSerializer.Serialize( movie ) );
-        await client.DeleteAsync( $"media/{movie.MediaInfo.Id}" );
+        foreach (var notFoundMovie in notFoundMovies)
+        {
+            await jellyseerrClient.DeleteAsync( $"media/{notFoundMovie.MediaId}" );
+        }
     }
+
 } );
 
 app.MapPost( "/radarr/notification", async ( [FromServices] IHttpClientFactory httpClientFactory, [FromBody] RadarrNotificationPayload payload ) =>
@@ -35,10 +88,10 @@ app.MapPost( "/radarr/notification", async ( [FromServices] IHttpClientFactory h
     Console.WriteLine( "Processing Radarr Notification" );
     Console.WriteLine( System.Text.Json.JsonSerializer.Serialize( payload ) );
 
-    if (payload.EventType.Equals( "MovieFileDelete", StringComparison.InvariantCultureIgnoreCase ))
+    if (payload.EventType.Equals( "MovieFileDelete", StringComparison.InvariantCultureIgnoreCase ) && !payload.DeleteReason.Equals( "upgrade", StringComparison.InvariantCultureIgnoreCase ))
     {
         Console.WriteLine( "Processing MovieFileDelete" );
-        var client = httpClientFactory.CreateClient( "JellySeerr" );
+        var client = httpClientFactory.CreateClient( "Jellyseerr" );
         var movie = await client.GetFromJsonAsync<JellySeerrMovie>( $"movie/{payload.Movie.TmdbId}" );
         await client.DeleteAsync( $"media/{movie.MediaInfo.Id}" );
     };
@@ -49,28 +102,39 @@ app.MapPost( "/sonarr/notification", async ( [FromServices] IHttpClientFactory h
     Console.WriteLine( "Processing Sonarr Notification" );
     Console.WriteLine( System.Text.Json.JsonSerializer.Serialize( payload ) );
 
-    if (payload.EventType.Equals( "SeriesDelete", StringComparison.InvariantCultureIgnoreCase ))
+    if (payload.EventType.Equals( "SeriesDelete", StringComparison.InvariantCultureIgnoreCase ) && !payload.DeleteReason.Equals( "upgrade", StringComparison.InvariantCultureIgnoreCase ))
     {
         Console.WriteLine( "Processing SeriesDelete" );
-        var client = httpClientFactory.CreateClient( "JellySeerr" );
+        var client = httpClientFactory.CreateClient( "Jellyseerr" );
 
     };
 
-    if (payload.EventType.Equals( "EpisodeFileDelete", StringComparison.InvariantCultureIgnoreCase ))
+    if (payload.EventType.Equals( "EpisodeFileDelete", StringComparison.InvariantCultureIgnoreCase ) && !payload.DeleteReason.Equals( "upgrade", StringComparison.InvariantCultureIgnoreCase ))
     {
         Console.WriteLine( "Processing EpisodeFileDelete" );
-        var client = httpClientFactory.CreateClient( "JellySeerr" );
+        var client = httpClientFactory.CreateClient( "Jellyseerr" );
 
     };
 } );
 
 app.Run();
 
+public class JellyfinSearchResult
+{
+    public List<JellyfinItem> Items { get; set; }
+}
+public class JellyfinItem
+{
+    public string Id { get; set; }
+    public string Name { get; set; }
+
+}
+
 public class JellySeerrMovie
 {
     public int Id { get; set; }
 
-    public JellySeerrMedia MediaInfo { get; set; }
+    public JellyseerrMedia MediaInfo { get; set; }
 
 }
 
@@ -80,12 +144,21 @@ public class JellySeerrTv
 
 }
 
-public class JellySeerrMedia
+public class JellyseerrMedia
 {
     public int Id { get; set; }
+    public int TmdbId { get; set; }
+    public string MediaType { get; set; }
+
+    public string JellyfinMediaId { get; set; }
+
+    public string JellyfinMediaId4k { get; set; }
 
 }
-
+public class JellyseerrMediaSearchResult
+{
+    public List<JellyseerrMedia> Results { get; set; }
+}
 public class SonarrNotificationPayload
 {
     public string EventType { get; set; }
